+++ public/index.php
@@ -63,6 +63,12 @@ require __DIR__ . '/../src/app/vulnerabilities/Forced Browsing Auth Bypass.php';
 require __DIR__ . '/../src/app/vulnerabilities/Parameter Modification Auth Bypass.php';
 require __DIR__ . '/../src/app/vulnerabilities/Session ID Modification Auth Bypass.php';
 require __DIR__ . '/../src/app/vulnerabilities/Spring4shell.php';
+require __DIR__ . '/../src/app/vulnerabilities/FP/X-Powered-By Header.php';
+require __DIR__ . '/../src/app/vulnerabilities/FP/Dangerous JS Functions.php';
+require __DIR__ . '/../src/app/vulnerabilities/GitCredentialsDisclosure.php';
+require __DIR__ . '/../src/app/vulnerabilities/WebDAVRoutes.php';
+require __DIR__ . '/../src/app/vulnerabilities/CockroachDBBrokenAccess.php';
+require __DIR__ . '/../src/app/vulnerabilities/SymfonySecretFragment.php';
 require __DIR__ . '/../src/app/vulnerabilities/CockroachDBBrokenAccess.php';
 require __DIR__ . '/../src/app/vulnerabilities/SymfonySecretFragment.php';
 require __DIR__ . '/../src/app/vulnerabilities/WebDAVRoutes.php';
@@ -74,7 +80,6 @@ require __DIR__ . '/../src/app/vulnerabilities/Atlassian Confluence Information 
 require __DIR__ . '/../src/app/vulnerabilities/Nginx Merge Slashes Path Traversal.php';
 require __DIR__ . '/../src/app/vulnerabilities/FP/X-Powered-By Header.php';
 require __DIR__ . '/../src/app/vulnerabilities/FP/Dangerous JS Functions.php';
-
 
 // Define Custom Error Handler
 $customErrorHandler = function (